backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1

# LlamaIndex - Get latest versions for Python 3.13 support
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-weaviate

# Database
weaviate-client>=3.26.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.2.0
websockets>=12.0
python-multipart